07_한글_PPT를_영문_PPT로_변경하기/requirements.txt
python-pptx
googletrans==4.0.0-rc1
